Refactor HostControls with functional components and hooks.
