+++ src/client/hostControls.tsx
@@ -1,182 +1,173 @@
-import React, { Component } from "react";
+import React from "react";
+import { SetParamType } from "../common/messages";
 import { t } from "./translation_ru";
+import { useDebouncedCallback } from 'use-debounce';
 
-export class HostControls extends Component<{
-    data: FullState,
-    socket: WebSocketChannel,
-    refreshState: () => void
-}> {
-    debouncedEmitTimer?: ReturnType<typeof setTimeout>;
+type GameSettingType = {
+    param: SetParamType,
+    label: string,
+    icon: string,
+    min?: number,
+    max?: number
+}
 
-    debouncedEmit(messageType: string, paramType: string, value: number) {
-        if (typeof this.debouncedEmitTimer !== 'undefined') {
-            clearTimeout(this.debouncedEmitTimer);
+const gameSettings: GameSettingType[] = [
+    { param: 'playerTime', label: 'player time', icon: 'alarm', min: 0},
+    { param: 'teamTime', label: 'team time', icon: 'alarm', min: 0},
+    { param: 'masterTime', label: 'master time', icon: 'alarm_on', min: 0},
+    { param: 'revealTime', label: 'reveal time', icon: 'alarm_on', min: 0},
+    { param: 'wordsLevel', label: 'words level', icon: 'school', min: 1, max: 4},
+    { param: 'goal', label: 'goal', icon: 'flag', min: 1}
+];
+
+type SettingInputProps = FullProps & { setting: GameSettingType };
+
+const SettingInput = ({data, socket, setting}: SettingInputProps) => {
+    const { param, label, icon, min, max } = setting;
+    //generate className from label e.g. "set-player-time"
+    const className = ['set', ...label.split(' ')].join('-');
+    const { hostId, userId, phase, paused } = data;
+    const isHost = hostId === userId;
+    const inProcess = phase !== 0 && !paused;
+    const editable = isHost && !inProcess;
+    const value = data[param];
+
+    const [update] = useDebouncedCallback((newVal: number) => {
+        if (!isNaN(newVal)) {
+            socket.emit("set-param", param, newVal);
         }
-        this.debouncedEmitTimer = setTimeout(() => {
-            this.props.socket.emit(messageType, paramType, value)
-        }, 100);
-    }
+    }, 100);
 
-    changeParam(value: number, type: string) {
-        this.debouncedEmit("set-param", type, value);
-    }
+    return (
+        <div className={className}>
+            <i title={t(label)} className="material-icons">
+                {icon}
+            </i>
+            {(editable) ? (
+                <input
+                    id={className}
+                    type="number"
+                    defaultValue={value}
+                    min={min}
+                    max={max}
+                    onChange={evt => update(evt.target.valueAsNumber)}
+                />
+            ) : (
+                <span className="value">
+                    {value}
+                </span>
+            )}
+        </div>
+    )
+}
 
-    clickChangeName() {
-        const { data, socket } = this.props;
-        const { playerNames, userId } = data;
-        popup.prompt({content: "New name", value: playerNames[userId] || ""}, (evt) => {
-            if (evt.proceed && evt.input_value && evt.input_value.trim()) {
-                socket.emit("change-name", evt.input_value.trim());
-                localStorage.userName = evt.input_value.trim();
-            }
+const GameSettings = ( {data, socket}: FullProps ) => (
+    <div className="host-controls-menu">
+        <div className="little-controls">
+            <div className="game-settings">
+                {gameSettings.map((setting) => (
+                    <SettingInput data={data} socket={socket} setting={setting} />
+                ))}
+            </div>
+        </div>
+    </div>
+)
+
+
+type HostControlsProps = FullProps & { refreshState: () => void };
+
+type ButtonProps = {
+    icon: string;
+    onClick: () => void;
+}
+
+const SideButtons = ( {data, socket, refreshState }: HostControlsProps ) => {
+    const { hostId, userId, phase, paused, teamsLocked, timed, playerNames } = data;
+    const isHost = hostId === userId;
+    const inProcess = phase !== 0 && !paused;
+    const buttons: ButtonProps[] = [];
+    if (isHost) {
+        buttons.push({
+            icon: 'store',
+            onClick: () => socket.emit("set-room-mode", false)
         });
+        buttons.push({
+            icon: (inProcess) ? 'pause' : 'play_arrow',
+            onClick: () => socket.emit("toggle-pause")
+        });
+        if (paused) {
+            buttons.push({
+                icon: (teamsLocked) ? 'lock_outline' : 'lock_open',
+                onClick: () => socket.emit("toggle-lock")
+            });
+            buttons.push({
+                icon: (timed) ? 'alarm' : 'alarm_off',
+                onClick: () => socket.emit("toggle-timed")
+            });
+            buttons.push({
+                icon: 'sync',
+                onClick: () => popup.confirm(
+                    {content: "Restart? Are you sure?"},
+                    (evt) => evt.proceed && socket.emit("restart")
+                )
+            });
+        }
     }
 
-    toggleTheme() {
+    const changeName = () => popup.prompt({
+        content: "New name",
+        value: playerNames[userId] || ""
+    }, ({proceed, input_value}) => {
+        const newName =  input_value?.trim();
+        if (proceed && newName) {
+            socket.emit("change-name", newName);
+            localStorage.userName = newName;
+        }
+    })
+    buttons.push({
+        icon: 'edit',
+        onClick: changeName
+    });
+
+    const toggleMuteSounds = () => {
+        localStorage.muteSounds = !parseInt(localStorage.muteSounds) ? 1 : 0;
+        refreshState();
+    };
+    buttons.push({
+        icon: (parseInt(localStorage.muteSounds)) ? 'volume_off' : 'volume_up', 
+        onClick: toggleMuteSounds
+    });
+
+    const toggleTheme = () => {
         localStorage.darkThemeDixit = !parseInt(localStorage.darkThemeDixit) ? 1 : 0;
         document.body.classList.toggle("dark-theme");
-        this.props.refreshState()
-    }
+        refreshState();
+    };
+    buttons.push({
+        icon: (parseInt(localStorage.darkThemeDixit)) ? 'wb_sunny' : 'brightness_2', 
+        onClick: toggleTheme
+    });
 
-    toggleMuteSounds() {
-        localStorage.muteSounds = !parseInt(localStorage.muteSounds) ? 1 : 0;
-        this.props.refreshState()
-    }
-
-    clickTogglePause() {
-        this.props.socket.emit("toggle-pause");
-    }
-
-    toggleTeamLockClick() {
-        this.props.socket.emit("toggle-lock");
-    }
-
-    clickRestart() {
-        popup.confirm(
-            {content: "Restart? Are you sure?"},
-            (evt) => evt.proceed && this.props.socket.emit("restart")
-        );
-    }
-
-    toggleTimed() {
-        this.props.socket.emit("toggle-timed");
-    }
-
-    setRoomMode() {
-        this.props.socket.emit("set-room-mode", false);
-    }
-
-
-    render() {
-        const
-            data = this.props.data,
-            isHost = data.hostId === data.userId,
-            inProcess = data.phase !== 0 && !data.paused;
-        return (
-            <div className="host-controls" onTouchStart={(e) => (e.target as HTMLElement).focus()}>
-                {data.timed ? (<div className="host-controls-menu">
-                    <div className="little-controls">
-                        <div className="game-settings">
-                            <div className="set-player-time"><i title={t("player time")}
-                                                                className="material-icons">alarm</i>
-                                {(isHost && !inProcess) ? (<input id="player-time"
-                                                                type="number"
-                                                                defaultValue={data.playerTime}
-                                                                min="0"
-                                                                onChange={evt => !isNaN(evt.target.valueAsNumber)
-                                                                    && this.changeParam(evt.target.valueAsNumber, "playerTime")}
-                                />) : (<span className="value">{data.playerTime}</span>)}
-                            </div>
-                            <div className="set-team-time"><i title={t("team time")}
-                                                            className="material-icons">alarm</i>
-                                {(isHost && !inProcess) ? (<input id="team-time"
-                                                                type="number"
-                                                                defaultValue={data.teamTime} min="0"
-                                                                onChange={evt => !isNaN(evt.target.valueAsNumber)
-                                                                    && this.changeParam(evt.target.valueAsNumber, "teamTime")}
-                                />) : (<span className="value">{data.teamTime}</span>)}
-                            </div>
-                            <div className="set-master-time"><i title={t("master time")}
-                                                                className="material-icons">alarm_on</i>
-                                {(isHost && !inProcess) ? (<input id="master-time"
-                                                                type="number"
-                                                                defaultValue={data.masterTime}
-                                                                min="0"
-                                                                onChange={evt => !isNaN(evt.target.valueAsNumber)
-                                                                    && this.changeParam(evt.target.valueAsNumber, "masterTime")}
-                                />) : (<span className="value">{data.masterTime}</span>)}
-                                <div className="set-reveal-time"><i title={t("reveal time")}
-                                                                    className="material-icons">alarm_on</i>
-                                    {(isHost && !inProcess) ? (<input id="reveal-time"
-                                                                    type="number"
-                                                                    defaultValue={data.revealTime}
-                                                                    min="0"
-                                                                    onChange={evt => !isNaN(evt.target.valueAsNumber)
-                                                                        && this.changeParam(evt.target.valueAsNumber, "revealTime")}
-                                    />) : (<span className="value">{data.revealTime}</span>)}
-                                </div>
-                                <div className="set-words-level"><i title={t("words level")}
-                                                                    className="material-icons">school</i>
-                                    {(isHost && !inProcess) ? (<input id="words-level"
-                                                                    type="number"
-                                                                    defaultValue={data.wordsLevel}
-                                                                    min="1"
-                                                                    max="4"
-                                                                    onChange={evt => !isNaN(evt.target.valueAsNumber)
-                                                                        && this.changeParam(evt.target.valueAsNumber, "wordsLevel")}
-                                    />) : (<span className="value">{data.wordsLevel}</span>)}
-                                </div>
-                                <div className="set-goal"><i title={t("goal")}
-                                                            className="material-icons">flag</i>
-                                    {(isHost && !inProcess) ? (<input id="goal"
-                                                                    type="number"
-                                                                    defaultValue={data.goal}
-                                                                    min="1"
-                                                                    onChange={evt => !isNaN(evt.target.valueAsNumber)
-                                                                        && this.changeParam(evt.target.valueAsNumber, "goal")}
-                                    />) : (<span className="value">{data.goal}</span>)}
-                                </div>
-                            </div>
-                        </div>
-                    </div>
-                </div>) : ""}
-                <div className="side-buttons">
-                    {data.userId === data.hostId ?
-                        <i onClick={() => this.setRoomMode()}
-                        className="material-icons exit settings-button">store</i> : ""}
-                    {isHost ? (!inProcess
-                        ? (<i onClick={() => this.clickTogglePause()}
-                            className="material-icons start-game settings-button">play_arrow</i>)
-                        : (<i onClick={() => this.clickTogglePause()}
-                            className="material-icons start-game settings-button">pause</i>)) : ""}
-                    {(isHost && data.paused) ? (data.teamsLocked
-                        ? (<i onClick={() => this.toggleTeamLockClick()}
-                            className="material-icons start-game settings-button">lock_outline</i>)
-                        : (<i onClick={() => this.toggleTeamLockClick()}
-                            className="material-icons start-game settings-button">lock_open</i>)) : ""}
-                    {(isHost && data.paused) ? (!data.timed
-                        ? (<i onClick={() => this.toggleTimed()}
-                            className="material-icons start-game settings-button">alarm_off</i>)
-                        : (<i onClick={() => this.toggleTimed()}
-                            className="material-icons start-game settings-button">alarm</i>)) : ""}
-                    {(isHost && data.paused)
-                        ? (<i onClick={() => this.clickRestart()}
-                            className="toggle-theme material-icons settings-button">sync</i>) : ""}
-                    <i onClick={() => this.clickChangeName()}
-                    className="toggle-theme material-icons settings-button">edit</i>
-                    {!parseInt(localStorage.muteSounds)
-                        ? (<i onClick={() => this.toggleMuteSounds()}
-                            className="toggle-theme material-icons settings-button">volume_up</i>)
-                        : (<i onClick={() => this.toggleMuteSounds()}
-                            className="toggle-theme material-icons settings-button">volume_off</i>)}
-                    {!parseInt(localStorage.darkThemeDixit)
-                        ? (<i onClick={() => this.toggleTheme()}
-                            className="toggle-theme material-icons settings-button">brightness_2</i>)
-                        : (<i onClick={() => this.toggleTheme()}
-                            className="toggle-theme material-icons settings-button">wb_sunny</i>)}
-                </div>
-                <i className="settings-hover-button material-icons">settings</i>
-            </div>
-        )
-    }
+    return (
+        <div className="side-buttons">
+            {buttons.map(({onClick, icon}) => (
+                <i onClick={onClick} className="material-icons settings-button">
+                    {icon}
+                </i>
+            ))}
+        </div>
+    )
 }
+
+export const HostControls = ( { data, socket, refreshState }: HostControlsProps) => (
+    <div
+        className="host-controls"
+        onTouchStart={(e) => (e.target as HTMLElement).focus()}
+    >
+        {data.timed && <GameSettings data={ data } socket={ socket } />}
+        <SideButtons data={ data } socket={ socket } refreshState={refreshState} />
+        <i className="settings-hover-button material-icons">
+            settings
+        </i>
+    </div>
+)
